+++ docs/changes.rst
@@ -16,6 +16,7 @@ Weblate 5.17.1
 .. rubric:: Bug fixes
 
 * :ref:`check-punctuation-spacing` check no longer triggers false positives for placeholders.
+* Client-side popup notifications triggered by JavaScript now use Bootstrap toasts.
 
 .. rubric:: Compatibility